colors fix

+++ src/main/java/com/hardsurf/wardrober/models/wardrobe/RgbColor.java
@@ -3,37 +3,37 @@ package com.hardsurf.wardrober.models.wardrobe;
 import java.io.Serializable;
 
 public class RgbColor implements Serializable {
-    private Byte red;
-    private Byte green;
-    private Byte blue;
+    private Integer red;
+    private Integer green;
+    private Integer blue;
 
-    public RgbColor(Byte red, Byte green, Byte blue) {
+    public RgbColor(Integer red, Integer green, Integer blue) {
         this.red = red;
         this.green = green;
         this.blue = blue;
     }
 
-    public Byte getRed() {
+    public Integer getRed() {
         return red;
     }
 
-    public void setRed(Byte red) {
+    public void setRed(Integer red) {
         this.red = red;
     }
 
-    public Byte getGreen() {
+    public Integer getGreen() {
         return green;
     }
 
-    public void setGreen(Byte green) {
+    public void setGreen(Integer green) {
         this.green = green;
     }
 
-    public Byte getBlue() {
+    public Integer getBlue() {
         return blue;
     }
 
-    public void setBlue(Byte blue) {
+    public void setBlue(Integer blue) {
         this.blue = blue;
     }
 }
